added MyParcelPageWithoutLogOutTest, HeaderPage and refactored HomePage, HomePageTest, MyParcelPage and MyParcelPageTest
