+++ src/main/java/pages/ForgotPasswordPage.java
@@ -4,7 +4,7 @@ import org.openqa.selenium.By;
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.WebElement;
 
-public class ForgotPasswordPage extends BasePage{
+public class ForgotPasswordPage extends BasePage {
     public ForgotPasswordPage(WebDriver driver) {
         super(driver);}
