+++ src/main/java/pages/SettingsPage.java
@@ -3,7 +3,6 @@ package pages;
 import org.openqa.selenium.By;
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.WebElement;
-import org.openqa.selenium.WebDriver;
 
 public class SettingsPage extends BasePage {
     public SettingsPage(WebDriver driver) {
