+++ src/main/java/pages/SignupPage.java
@@ -2,7 +2,7 @@ package pages;
 
 import org.openqa.selenium.WebDriver;
 
-public class SignupPage extends BasePage{
+public class SignupPage extends BasePage {
     public SignupPage(WebDriver driver) {
         super(driver);
     }
