+++ src/main/java/pages/SrchFldPage.java
@@ -2,8 +2,6 @@ package pages;
 
 import org.openqa.selenium.By;
 import org.openqa.selenium.WebDriver;
-import org.openqa.selenium.WebElement;
-import pages.BasePage;
 
 public class SrchFldPage extends BasePage {
     public SrchFldPage(WebDriver driver) {
