+++ src/test/java/tests/AuthorizedTest.java
@@ -1,5 +1,6 @@
 package tests;
 
+import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import pages.*;
 
@@ -8,16 +9,16 @@ import java.util.Map;
 import static utils.YAMLDeserializer.fromFileToMap;
 
 public abstract class AuthorizedTest extends BaseTest {
-
     protected Map<String, String> user;
-
     protected MyParcelsPage myParcelsPage;
+    protected HeaderPage headerPage;
 
     public AuthorizedTest() {
         super();
 
         user = fromFileToMap("user.yaml");
         myParcelsPage = new MyParcelsPage(driver);
+        headerPage = new HeaderPage(driver);
     }
 
     public AuthorizedTest(String file) {
@@ -30,12 +31,17 @@ public abstract class AuthorizedTest extends BaseTest {
     @BeforeEach
     public void beforeEach() {
         super.beforeEach();
-
-        myParcelsPage = homePage.clkMenuBtn()
+        myParcelsPage = headerPage.clkMenuBtn()
                 .clkLoginBtn()
                 .insertLoginFld(user.get("name"), user.get("password"))
                 .clkSubmitLogin();
+    }
 
+    @AfterEach
+    public void tearDown() {
+        /*headerPage.clkMenuBtn()
+                .clkUserLogOutBtn();*/
+        super.tearDown();
     }
 
 }
