+++ src/test/java/tests/home_page/HomePageTest.java
@@ -1,20 +1,20 @@
 package tests.home_page;
 
 import org.assertj.core.api.SoftAssertions;
-import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import pages.HomePage;
 import pages.MenuPage;
-import tests.BaseTest;
-import utils.DriverConfiguration;
+import tests.AuthorizedTest;
 
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
-public class HomePageTest extends BaseTest {
+public class HomePageTest extends AuthorizedTest {
+    protected HomePage homePage;
     protected MenuPage menuPage;
 
     public HomePageTest() {
         super();
+        homePage = new HomePage(driver);
         menuPage = new MenuPage(driver);
     }
 
@@ -22,16 +22,6 @@ public class HomePageTest extends BaseTest {
     public void verifyThatAllElementsAreDisplayed() {
         SoftAssertions softAssertions = new SoftAssertions();
 
-        softAssertions.assertThat(homePage.isTbrDsp())
-                .withFailMessage("Tbr isn't displayed").isTrue();
-        softAssertions.assertThat(homePage.isBtnMenuDsp())
-                .withFailMessage("BtnMenu isn't displayed").isTrue();
-        softAssertions.assertThat(homePage.isImgLogoDsp())
-                .withFailMessage("ImgLogo isn't displayed").isTrue();
-        softAssertions.assertThat(homePage.isLnkLogoDsp())
-                .withFailMessage("LnkLogo isn't displayed").isTrue();
-        softAssertions.assertThat(homePage.isBtnListBoxDsp())
-                .withFailMessage("BtnListBox isn't displayed").isTrue();
         softAssertions.assertThat(homePage.isHdngFindYourParcelDsp())
                 .withFailMessage("HeadingFindYourParcel isn't displayed").isTrue();
         softAssertions.assertThat(homePage.isInptFieldDsp())
@@ -69,7 +59,7 @@ public class HomePageTest extends BaseTest {
      */
     @Test
     public void verifyThatUserIsAbleToClkMenuBtnToOpen() {
-        homePage.clkMenuBtn();
+        headerPage.clkMenuBtn();
         assertTrue(menuPage.isTracksterHdngDsp(), "Menu button doesn't work correctly");
     }
 
@@ -79,7 +69,7 @@ public class HomePageTest extends BaseTest {
      */
     @Test
     public void verifyThatChangingEngToUaIsPossible() {
-        homePage.clkListBoxBtn()
+        headerPage.clkListBoxBtn()
                 .clkUaBtn();
         assertTrue(homePage.isUaHdngDisplayed(), "Language isn't correct");
     }
@@ -90,7 +80,7 @@ public class HomePageTest extends BaseTest {
      */
     @Test
     public void verifyThatChangingUaToEngIsPossible() {
-        homePage.clkListBoxBtn()
+        headerPage.clkListBoxBtn()
                 .clkUaBtn()
                 .clkListBoxBtn()
                 .clkEngBtn();
