+++ src/test/resources/app.properties
@@ -1,4 +1,4 @@
-webdriver.chrome.driver=C:\\WebDriver\\bin\\chromedriver.exe
+webdriver.chrome.driver=E:\\chromedriver_win32\\chromedriver.exe
 base.url=https://ttrackster.herokuapp.com/
 find.url=http://ttrackster.herokuapp.com/tracking?number=12345678
 user.email=fortestsnastia@gmail.com
